+++ demo/routes/index.js
@@ -1,6 +1,6 @@
 const Foxify = require('../../framework')
 
-let routes = new Foxify.Route()
+const routes = new Foxify.Router()
 
 routes.get('/', (req, res) => {
   res.render('index', {
@@ -9,7 +9,7 @@ routes.get('/', (req, res) => {
   })
 })
 
-routes.get('/greet/', (req, res) => {
+routes.get('/greet', (req, res) => {
   res.json({
     hello: 'world'
   })
@@ -37,7 +37,7 @@ routes.get('/greet-fast', {
 })
 
 routes.get('/404', (req, res) => {
-  throw new HttpException(404)
+  throw new HttpException("This is a demo", 404)
 })
 
 routes.get('/error', async (req, res) => {