+++ src/routing.old/httpMethods.ts
@@ -1,0 +1,3 @@
+import * as httpMethods from "methods";
+
+export default httpMethods.map((method) => method.toUpperCase());