+++ src/routing/httpMethods.ts
@@ -1,3 +1,11 @@
-import * as httpMethods from "methods";
+import * as http from "http";
 
-export default httpMethods.map((method) => method.toUpperCase());
+export type Method = "ACL" | "BIND" | "CHECKOUT" | "CONNECT" | "COPY" |
+  "DELETE" | "GET" | "HEAD" | "LINK" | "LOCK" | "M-SEARCH" | "MERGE" | "MKACTIVITY" |
+  "MKCALENDAR" | "MKCOL" | "MOVE" | "NOTIFY" | "OPTIONS" | "PATCH" | "POST" |
+  "PROPFIND" | "PROPPATCH" | "PURGE" | "PUT" | "REBIND" | "REPORT" | "SEARCH" |
+  "SOURCE" | "SUBSCRIBE" | "TRACE" | "UNBIND" | "UNLINK" | "UNLOCK" | "UNSUBSCRIBE";
+
+const METHODS = http.METHODS as Method[];
+
+export default METHODS;