+++ src/routing/index.ts
@@ -1,9 +1,9 @@
 import httpMethods from "./httpMethods";
+import * as Layer from "./Layer";
 import * as Router from "./Router";
-import * as Route from "./Route";
 
 export {
   httpMethods,
+  Layer,
   Router,
-  Route,
 };